Let action removeBlocks handle prompts and confirmations
