+++ packages/block-editor/src/utils/index.js
@@ -1,4 +1,4 @@
 export { default as transformStyles } from './transform-styles';
 export * from './block-variation-transforms';
 export { default as getPxFromCssUnit } from './parse-css-unit-to-px';
-export { BlockRemovalWarningModal } from './show-block-removal-warning';
+export { BlockRemovalWarningModal } from './block-removal';
